typed cast return

+++ 14-binary_tree_balance.c
@@ -15,5 +15,5 @@ int binary_tree_balance(const binary_tree_t *tree)
 		return (0);
 	_height_r = tree->right ? 1 + binary_tree_balance(tree->right) : 0;
 	_height_l = tree->left ? 1 + binary_tree_balance(tree->left) : 0;
-	return (_height_l - _height_r);
+	return ((int)(_height_l - _height_r));
 }
